requirements.txt
alembic==1.17.1
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.129.0
greenlet==3.3.2
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg[binary]==3.2.12
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
PyMuPDF==1.27.1
pyparsing==3.3.2
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.22
redis==5.2.1
requests==2.32.5
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
